// packages/types/src/userHelperKeys.ts
export enum UserHelperKeys {
  // QUICK START GUIDE
  SEND_TO_SALSES_MODAL_FEEDBACK_STEP = 'SEND_TO_SALSES_MODAL_FEEDBACK_STEP',
  CHECK_OUT_YOUR_TARGET_MARKETS = 'CHECK_OUT_YOUR_TARGET_MARKETS',
  CHECK_OUT_YOUR_BUYER_PERSONAS = 'CHECK_OUT_YOUR_BUYER_PERSONAS',
  CREATE_YOUR_FIRST_CADENCE_STEP = 'CREATE_YOUR_FIRST_CADENCE_STEP',
  START_TASK_FROM_CADENCE = 'START_TASK_FROM_CADENCE',
  SEND_FIRST_AUTO_EMAIL = 'SEND_FIRST_AUTO_EMAIL',
  SAVE_NUMBER_SETTINGS = 'SAVE_NUMBER_SETTINGS',
  CREATE_YOUR_FIRST_PITCH = 'CREATE_YOUR_FIRST_PITCH',
  MARK_AS_DONE_ATTEMPT = 'MARK_AS_DONE_ATTEMPT',
  TAKE_TOUR_ON_OUTBOX = 'TAKE_TOUR_ON_OUTBOX',
  // ADMIN QUICK START GUIDE
  CREATE_FIRST_TARGET_MARKET = 'CREATE_FIRST_TARGET_MARKET',
  ADD_FIRST_BUYER_PERSONA = 'ADD_FIRST_BUYER_PERSONA',
  DEFINE_FIRST_SCENARIO = 'DEFINE_FIRST_SCENARIO',
  SET_COMPANY_FIELDS_TOUR = 'SET_COMPANY_FIELDS_TOUR',
  SET_LEAD_FIELDS_TOUR = 'SET_LEAD_FIELDS_TOUR',
  TAKE_TOUR_ON_GENERAL_SETTINGS = 'TAKE_TOUR_ON_GENERAL_SETTINGS',
  CONNECT_CRM_TOUR = 'CONNECT_CRM_TOUR',
  DEFINE_QQ = 'DEFINE_QQ',
  SET_UP_DASHBOARDS_TOUR = 'SET_UP_DASHBOARDS_TOUR',
  INVITE_TEAM = 'INVITE_TEAM',
  // OTO QUICK START GUIDE
  CONNECT_SALESFORCE = 'CONNECT_SALESFORCE',
  CONNECT_EMAIL = 'CONNECT_EMAIL',
  SET_TIMEZONE = 'SET_TIMEZONE',
  START_USING_SALESFORCE_ON_TOP_OF = 'START_USING_SALESFORCE_ON_TOP_OF',

  //SHARED QSG KEYS
  SET_UP_REMINDERS = 'SET_UP_REMINDERS',
  CHOOSE_DIALER = 'CHOOSE_DIALER',
  SET_YOUR_EMAIL_SIGNATURE = 'SET_YOUR_EMAIL_SIGNATURE',
  CREATE_FIRST_COMPANY = 'CREATE_FIRST_COMPANY',
  CREATE_FIRST_LEAD = 'CREATE_FIRST_LEAD',
  LAUNCH_YOUR_FIRST_CADENCE = 'LAUNCH_YOUR_FIRST_CADENCE',
  TAKE_TOUR_PROSPECT_TAB = 'TAKE_TOUR_PROSPECT_TAB',
  CALL_AND_REPORT_RESULT = 'CALL_AND_REPORT_RESULT',
  CREATE_LEAD_FROM_LINKEDIN = 'CREATE_LEAD_FROM_LINKEDIN',
  LINK_FIRST_MESSAGE_LINKEDIN = 'LINK_FIRST_MESSAGE_LINKEDIN',
  TAKE_TOUR_ON_INBOX = 'TAKE_TOUR_ON_INBOX',
  DOWNLOAD_CHROME_EXTENSION = 'DOWNLOAD_CHROME_EXTENSION',
  CREATE_FIRST_EMAIL_TEMPLATE = 'CREATE_FIRST_EMAIL_TEMPLATE',
  CREATE_YOUR_FIRST_CADENCE = 'CREATE_YOUR_FIRST_CADENCE',
  CONNECT_EMAIL_ACCOUNT = 'CONNECT_EMAIL_ACCOUNT',
  ENABLE_KPI_METRICS = 'ENABLE_KPI_METRICS',
  CREATE_FIRST_LIST = 'CREATE_FIRST_LIST',
  SEND_YOUR_FIRST_EMAIL = 'SEND_YOUR_FIRST_EMAIL',

  // REST
  COMPLETE_WELCOME_SCREEN = 'COMPLETE_WELCOME_SCREEN',
  SEEN_INACTIVE_DISCOVERY_TOOLTIP = 'SEEN_INACTIVE_DISCOVERY_TOOLTIP',
  NEW_MEETING_MODAL = 'NEW_MEETING_MODAL',
  NEW_INVITEES_SYNC = 'NEW_INVITEES_SYNC',
  PREVIEW_EMAIL_MODAL = 'PREVIEW_EMAIL_MODAL',
  DONT_SHOW_AGAIN_CREATED_USER_MODAL = 'DONT_SHOW_AGAIN_CREATED_USER_MODAL',
  DONT_SHOW_AGAIN_MEETING_FUNNEL_DATE_HINT = 'DONT_SHOW_AGAIN_MEETING_FUNNEL_DATE_HINT',
  SCHEDULE_NEXT_STEP_DATETIME_FILTER = 'SCHEDULE_NEXT_STEP_DATETIME_FILTER',
  WIZARDS_RECALL_DATETIME = 'WIZARDS_RECALL_DATETIME',
  SHOW_DISABLED_VALUES_FILTERS_DASHBOARDS = 'SHOW_DISABLED_VALUES_FILTERS_DASHBOARDS',
  SHOW_DISABLED_VALUES_FILTERS_REPORTS = 'SHOW_DISABLED_VALUES_FILTERS_REPORTS',
}